price

+++ app/src/main/java/com/pg/personalaccounting/core/models/Account.kt
@@ -2,7 +2,12 @@ package com.pg.personalaccounting.core.models
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.util.*
 
 @Entity(tableName = "accounts")
-data class Account(@PrimaryKey(autoGenerate = true) val id:Int, val accType:String, val accNumber: String, val accDate:String, var balance:Float, val bankName:String)
+data class Account(
+    @PrimaryKey(autoGenerate = true) val id: Int, val accType: String,
+    val accNumber: String,
+    val accDate: String,
+    var balance: Double,
+    val bankName: String
+)
